Fix bugs that caused mutations with side effects.
Previously, performing replacement or substitution on an AST could
alter the AST that was being used as a replacement or as a substitution.
This meant that, for example, mutations could interfere with other
mutations that shared some of the same AST as their match.root.

This fix ensures that replacement and substitution do not use terms from
the original AST in the rewriting.